+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Commit messages follow [Conventional Commits](https://www.conventionalcommits.or
 
 ## [Unreleased]
 
+### Changed
+
+- Cold-start screen picks a random quote photo from the splash set
+
 ## [0.65.0] - 2026-08-17
 
 ### Changed